sqlite_to_postgres/requirements.txt
asgiref==3.8.1
Django==4.2.11
-e git+https://github.com/jazzband/django-debug-toolbar.git@8f4fa8e7f0a5ddc539f98fb2767b17fd568732d3#egg=django_debug_toolbar
django-split-settings==1.3.2
Faker==25.0.1
flake8==6.1.0
install==1.3.5
isort==5.13.2
mccabe==0.7.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
